backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.10.3
pydantic-settings==2.6.1
redis==5.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
email-validator==2.2.0
faker==33.1.0
